Cleanup time point.

+++ source/common/time/timer.h
@@ -10,23 +10,23 @@ public:
         m_start = std::chrono::high_resolution_clock::now();
     }
 
-    inline float time()
+    inline double time()
     {
         const auto now       = std::chrono::high_resolution_clock::now();
-        const auto time_span = std::chrono::duration_cast<std::chrono::duration<float>>(now - m_start);
+        const auto time_span = std::chrono::duration_cast<std::chrono::duration<double>>(now - m_start);
         return time_span.count();
     }
 
-    inline float interval()
+    inline double interval()
     {
-        const float t = time();
+        const double t = time();
         start();
         return t;
     }
 
-    inline float intervalSmooth()
+    inline double intervalSmooth()
     {
-        static constexpr float kAlpha = 0.1f;
+        static constexpr double kAlpha = 0.1f;
 
         m_interval = (kAlpha * interval()) + (1.0f - kAlpha) * m_interval;
         return m_interval;
@@ -35,6 +35,6 @@ public:
 private:
     std::chrono::high_resolution_clock::time_point m_start;
 
-    float m_interval = 0.0f;
+    double m_interval = 0.0f;
 };
 } // namespace Tyr::Common
